knocked off a couple of FIXMEs in vncDesktopThread::run_undetached():
// FIXME: rewrite this stuff to eliminate code duplication (ses above).
// FIXME: Use time() instead of GetSystemTime().

i moved the timimg code to a new class called GracePeriod, and replaced
the duplicate code with calls to an instance of this class.

there's one bug right now: the time interval for the GracePeriod is only
set when vncDesktop::Init() is called. this needs to be updated when
the user changes the remote input override interval. i'll fix this asap.

i also fixed a couple minor issues in some ifdef'd HORIZONLIVE code.


git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@1943 3789f03b-4d11-0410-bbf8-ca57d06f2519