Make Xvnc build on modern distributions, such as Fedora Core 3:
- The configure script adds "LinuxCLibMajorVersion 6" to
platform.cf. This makes the build system define _XOPEN_SOURCE, which
causes __USE_XOPEN to be defined. This, in turn, makes fds_bits
available. (As far as I know, imake should find out and define
LinuxCLibMajorVersion, but for some reason, this doesn't work.)

- lib/Xau/AuRead.c and lib/Xau/AuWrite.c has been updated to the
latest X.Org versions. There are only small fixes: Not defining a
prototype for malloc, for example.


git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@2007 3789f03b-4d11-0410-bbf8-ca57d06f2519